backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
faiss-cpu
sentence-transformers
pydantic
pydantic-settings
python-multipart
sqlalchemy
bcrypt
python-jose[cryptography]
chromadb
numpy
aiofiles
